browser/components/resistfingerprinting/test/mochitest/test_iframe.html

<!doctype html>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
<body>
<script>
  add_task(async function() {
    await SpecialPowers.pushPrefEnv({
      "set": [["privacy.resistFingerprinting", true]],
    });
    is(screen.width, window.innerWidth, "Width should be spoofed");
    is(screen.height, window.innerHeight, "Height should be spoofed");
    let iframe = document.createElement("iframe");
    document.body.appendChild(iframe);
    is(iframe.contentWindow.screen.width, iframe.contentWindow.innerWidth, "Width should be spoofed in iframe");
    is(iframe.contentWindow.screen.height, iframe.contentWindow.innerHeight, "Height should be spoofed in iframe");
  });
</script>
</body>
